+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Package::signature_key_fingerprints()`
 - The ability to mark files as `%missingok`
 
+### Fixed
+
+- Added validation to reject control characters in inputs to `PackageBuilder`.
+- Added validation to match RPM's treatment of characters allowed or disallowed in package names.
+- Detect and strip control characters when parsing packages.
+
 ### Changed
 
 - Bump `pgp` to `0.19.0`